Reorganized test file structure.
